pyproject.toml
[project]
name = "fastmcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "fastmcp>=2.14.4",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
